Working Version 1
This is a working-ish version of the web app. You can add candidates and it will determine the winner, but there is not functionality when there is a tie or if there is only one candidate that has been submitted. You can only add votes at the time of candidate creation -fix this.  Will work on these features and make the design a lot more robust.