Mock-Election V2.0
This is a complete update to cosmetics and logic of the website. Changed how the submit and determine winner buttons are placed and scaled responsively. Added a border around the form.Changed the padding and placement of text in the header and colorbar(orange strip).  Fixed various logic issues when determining the winner. Multiple winners in various races will no longer display extra copies of winners in the winner display box. Made the site more mobile friendly -still some work to do on this.
Features for the future - Add votes after candidates and initial votes have been entered.